Moviepy

+++ highzer/clip.py
@@ -1,11 +1,14 @@
+import os
 from utils import pprint, locate_folder
 from twitch import get_top_clips
 from fpeg import concat, sample, merge, convert, concat_fast, cat
 import requests
+from movie import concat_clips
 
 
 def cut_clips(ident, period, game, channel):
     res = get_top_clips(period, game, channel)
+    print("Received TOP clips")
 
     # last clip is most viewed
     clips = sorted(res, key=lambda x: x["views"])
@@ -17,32 +20,31 @@ def cut_clips(ident, period, game, channel):
         raw_name = f"{folder}/{clip['slug']}"
         filename = f"{raw_name}.mp4"
         download_clip(clip, filename)
-        # sample(filename)
-
-        converted_name = f"{raw_name}.mpg"
-        convert(filename, converted_name)
-        filename = converted_name
 
         files.append(filename)
 
-    # cat(files, f"{folder}/clip_merged.mpeg")
-    # import os
-    # concat([os.path.abspath(file) for file in files], os.path.abspath(f"{folder}/clip_merged.mp4"), os.path.abspath(f"{folder}/temp"))
-    # concat(files, f"{folder}/clip_merged.mp4", f"{folder}/temp")
-    # merge(files, f"{folder}/clip_merge.mp4")
     out_file = f"{folder}/merged.mp4"
-    convert(concat_fast(files, f"{folder}/merged.mpg"), out_file)
+    concat_clips(files, out_file)
 
     return clips, out_file
 
 
-def download_clip(clip, filename):
+def download_clip(clip, filename, force=False):
+    if not force and os.path.isfile(filename):
+        print("Clip already downloaded.")
+        return
+
     thumb_url = clip["thumbnails"]["tiny"]
     mp4_url = thumb_url.split("-preview", 1)[0] + ".mp4"
 
+    print(f"Downloading clip: {filename}")
     res = requests.get(mp4_url)
     with open(filename, "wb") as file:
         file.write(res.content)
 
-    print(f"downloaded clip: {filename}")
+    print(f"Downloaded clip: {filename}")
     return filename
+
+
+if __name__ == '__main__':
+    cut_clips('testing', 'week', 'League of Legends', None)
